requirements.txt
# Web Frameworks
Flask==3.0.1

# AI and Google Integration
google-generativeai==0.5.2

# Utilities
python-dotenv==1.0.1